Move midirender methods from score to its own class
